// ecis-nodejs-zgwx-feat_240603/service/account-zgxw/ecissdk/model/data.ts
export interface MysqlSelResp {
  result: string
  data: { rows: any[] }
  msg?: string
}
export interface MysqlUpdateResp {
  result: string
  data: { affect: any }
  msg?: string
}
export interface MysqlInsertResp {
  result: string
  data: { affect: any }
  msg?: string
}

export interface MysqlDelResp {
  result: string
  data: { affect: any }
  msg?: string
}